Add a MessageChannel interface for bidirectional, asynchronous message-passing channels.
The interface is based on the API for goog.net.xpc.CrossDomainChannel, but also
implemented by a wrapper for goog.gears.Worker. This means that iframe messaging
and Gears worker messaging can be used interchangeably, and chained using
goog.messaging.pipe.

This is useful for e.g. allowing a Gears worker to talk to an iframe, using the
main page as a proxy.

R=pupius
DELTA=894  (892 added, 0 deleted, 2 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=204390


git-svn-id: https://closure-library.googlecode.com/svn/trunk@262 0b95b8e8-c90f-11de-9d4f-f947ee5921c8