Automated g4 rollback.
*** Reason for rollback ***

Broke builds

*** Original change description ***

Improve security of XPC native transport by having it check the origin of received messages against the PEER_HOSTNAME config attribute, if set.  If the origin of the message is known (which is only true in native transport) and the PEER_HOSTNAME param was set, the two are required to match, or the message will be ignored.

R=ssaviano
DELTA=63  (0 added, 57 deleted, 6 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=9


git-svn-id: https://closure-library.googlecode.com/svn/trunk@493 0b95b8e8-c90f-11de-9d4f-f947ee5921c8