Removes browser checks when adding ARIA roles, states, and properties to components. Also removes the setNoBrowserCheck option.
This was originally submitted, but rolled back because
it broke the build. The reason was that getAttribute on IE6 and IE7 return
boolean values when checking for ARIA state instead of string values. The
method in goog.dom.a11y that looks at ARIA states has been updated to account
for this quirk.

R=pupius
DELTA=173  (0 added, 101 deleted, 72 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=90


git-svn-id: https://closure-library.googlecode.com/svn/trunk@524 0b95b8e8-c90f-11de-9d4f-f947ee5921c8