The Menu controller renders the menu portion of the control at the end of the document rather than within the parent div. This causes problems with things like popups and autocomplete that check if MOUSEDOWN events happened within an element they own.  The result is that any click within a drop-down menu in a popup will cause the popup to close itself.
The easiest way to fix this behavior is to have the menu rendered within the same scope as the label...otherwise it is a complete hack to prevent autocomplete, or popups, or submenus from closing themselves when someone clicks on a menu.

R=gboyer
DELTA=51  (43 added, 4 deleted, 4 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=183


git-svn-id: https://closure-library.googlecode.com/svn/trunk@557 0b95b8e8-c90f-11de-9d4f-f947ee5921c8