A bug in JScript in IE < 8 causes asserts to throw an exception if DOM objects are passed in -- but only if there is no try, catch, or finally block in the stack when the assert is called
This CL fixes it by adding an empty catch block. This should not cause any difference in behavior for browsers that run Javascript correctly, because the return inside of the finally already stops the propagation of uncaught exceptions (I was surprised by this too, so I moved the return out of the finally because it's confusing and the catch block will now do the same thing).

The testcase this CL adds has to proxy the call through the VBScript engine; otherwise, the try block from the test framework will be in the callstack, preventing the JScript bug from ever occurring.

R=pupius
DELTA=20  (18 added, 0 deleted, 2 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=353


git-svn-id: https://closure-library.googlecode.com/svn/trunk@649 0b95b8e8-c90f-11de-9d4f-f947ee5921c8