Add mock FileSystem classes for testing.
The primary goal of this API is to allow clients of the FileSystem API to be
tested without touching the actual filesystem at all. The secondary goal is to
allow them to be tested even on browsers that don't support the API; to this
end, even objects like Blobs have been mocked out.

R=nicksantos
DELTA=2230  (2230 added, 0 deleted, 0 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=356


git-svn-id: https://closure-library.googlecode.com/svn/trunk@652 0b95b8e8-c90f-11de-9d4f-f947ee5921c8