Fixed iOS pre-4.2 roleselector.
There seems to be a bug in iOS prior to 4.2 where menubutton selection fails - instead the menu is just closed without any selection made. This change makes it so that the blur event isn't listened to on these devices, so that the selection event can be processed.

R=pallosp
DELTA=15  (14 added, 0 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2065


git-svn-id: https://closure-library.googlecode.com/svn/trunk@966 0b95b8e8-c90f-11de-9d4f-f947ee5921c8