Change to Closure labelinput widget to use ARIA labels or the placeholder attribute where it is supported.
Without this change, screen reader users who encounter input fields of this type do not see any label at all, because the label is deleted as soon as a screen reader focus on the input field. This is a big accessibility problem in many projects.

The HTML5 placeholder attribute is a solution to this problem - the browser does the text rendering inside the input field for us and raises that text to screen readers. So in browsers where the placeholder attribute is supported, it should be used.

In browsers where the placeholder attribute is not supported, we can use the ARIA state aria-label to specify to screen readers the label of the input field.

R=pupius
DELTA=116  (109 added, 1 deleted, 6 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2333


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1052 0b95b8e8-c90f-11de-9d4f-f947ee5921c8