Workaround behavior of goog.ui.Checkbox::toggle() which toggles undeterminate->checked, whereas we prefer undetermined->UNchecked.
The reason being that a common workflow is to select n items (checkbox is now in undetermined state), perform an action on them (like mark-as-read), and then to unselect these items with a single click on the checkbox.

R=chrishenry,nnaze
DELTA=34  (17 added, 16 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2696


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1140 0b95b8e8-c90f-11de-9d4f-f947ee5921c8