Small change to controlrenderer.js remove the aria-pressed state that is set whenever a control is ACTIVE. Setting the aria-pressed state this way is incorrect and causes strange screen reader behavior. And aria-pressed isn't even supposed to be used by most controls, only buttons.
For example, a checkbox is "active" while a user is clicking on it,
but not before or after. This means that aria-pressed="false" most of
the time, except while the user is clicking and then it is set to be
aria-pressed="true". This is not correct behavior - aria-pressed is not
supposed to describe transient pressed state.

This change also updates the test to support this new behavior. Also
changes confusing wording in checkboxrenderer.js.

R=nnaze
DELTA=58  (48 added, 1 deleted, 9 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2943


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1199 0b95b8e8-c90f-11de-9d4f-f947ee5921c8