Add a helpful comment to MockClock regarding its use with goog.fx.Animation.
Background:

We spent most of a day (or more) hunting down an issue where tests in the same file were failing due to an interaction between MockClock and Animations.

The symptom was that mockClock.tick(x) would not advance animations in any test function that executed *after* a test function that (a) installed a MockClock, (b) started an animation, and (c) did *not* complete that animation by calling tick().

This is because goog.fx.anim creates a single, global animation timeout via window.setTimeout, which, if MockClock is installed per (a), gets registered in that clock's internal queue.  If the clock is disposed of without firing that animation timeout, then the timeout will never be fired, even if new clocks are created.  All new animations will be registered with the defunct timeout, resulting in no animations.

One solution would be to call all registered timeouts in the MockClock's queue on uninstall(), but I don't know if that's reasonable or not.

R=pupius
DELTA=17  (17 added, 0 deleted, 0 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3608


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1366 0b95b8e8-c90f-11de-9d4f-f947ee5921c8