Automated g4 rollback.
*** Reason for rollback ***

This change breaks all hovercards

*** Original change description ***

Small update to the HoverCard to decouple the logic for finding trigger elements and the logic for rendering hovercards. Especially useful / necessary if iframes are in play. The update in advancedtooltip is to respect the document context of the anchor element, and the update in tooltip is to use the proper coordinates if iframed.

R=gboyer
DELTA=13  (0 added, 6 deleted, 7 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3641


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1371 0b95b8e8-c90f-11de-9d4f-f947ee5921c8