Updates closure slider component, such that when the move to point (on click) option is enabled, the component determines which value to move to based on the current step size.
The slider component currently disregards the step size in this case, which
allows users to position the thumb at arbitrary locations when move to point is
enabled. This is inconsistent with dragging the thumb, the latter restricts
thumb placement only to values based on step size. This inconsistency makes it
unfeasible to use the move to point option for sliders that have a step size.


R=chrishenry,arv
DELTA=87  (64 added, 4 deleted, 19 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4215


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1581 0b95b8e8-c90f-11de-9d4f-f947ee5921c8