Annotate some parameters as optional when they were already being treated as optional:
- The BrowserChannel constructor previously required a
clientVersion to be provided, but some users didn't provide
one, the client-side code only sets the CVER parameter when
clientVersion is non-null, and the server substitutes 1 when
CVER is not provided. 

- The BrowserChannel.connect function was previously declared
to require extraParams, but it correctly treated it as
optional, and not all users provide it.

Also annotate the BrowserChannel private state_ and
channelDebug_ fields as non-null.

R=mpd
DELTA=14  (2 added, 0 deleted, 12 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4272


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1614 0b95b8e8-c90f-11de-9d4f-f947ee5921c8