Automated g4 rollback
*** Reason for rollback ***

The new type is too loose for some project. It's breaking
their compilation. !Array.<*>, the * part is causing
compiler to suddenly understand that the values in this
array is * while previously it just couldn't infer it.

*** Original change description ***

Tighten return values of QueryData#getKeys and
#getValues. Also attempt to simplify getValues slightly
(and remove the lone early return). I assume this was
a remnant of IE6 days optimization.

R=nnaze,brenneman
DELTA=34  (14 added, 5 deleted, 15 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4342


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1645 0b95b8e8-c90f-11de-9d4f-f947ee5921c8