Creates goog.labs.structs.Map. This is meant to be a labs implementation of a possible replacement for goog.structs.Map and evolve a cleaner API.
At the end of the day, it is more likely that we will try to gut the
implementation of goog.structs.Map with whatever the implementation we
come up with in the labs version (or, we can decide to deprecate
goog.structs.Map and just use the new implementation).

R=arv
DELTA=847  (847 added, 0 deleted, 0 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4370


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1653 0b95b8e8-c90f-11de-9d4f-f947ee5921c8