Fixed 2 further flakiness.
With continuationtestcase_test, the flakiness is caused by:
1) testOutOfOrderWaits - the timeout handlers can be executed at the
same time in IE7/8, FF if the timeout is too small. So really, it
doesn't provide any guarantee that out of order waits will be called
in order of timeout.
2) testWaitForCondition - the handler may not be invoked every 10ms
as the test implies. Older browsers (IE7 in particular) may invoke them
at higher interval.

With deferredtestcase_test, it's not flakiness. It's a JS error that
happens after the test completes. This sometime caused the next test
to execute in the same browser to fail completely.

R=nicksantos
DELTA=39  (9 added, 12 deleted, 18 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4452


git-svn-id: https://closure-library.googlecode.com/svn/trunk@1700 0b95b8e8-c90f-11de-9d4f-f947ee5921c8