Automated g4 rollback
*** Reason for rollback ***

This change broke builds

*** Original change description ***

This CL changes the way properties are added to a mock such that IE can add base class properties (which are non-enumerable on IE).

Also fixes a bunch of test issues that showed up with the change. In these, Mocks were being created with a type that cannot be instantiated directly or tests were depending on the base Object prototype .toString() function in objects that had explicitly overridden it.


Revision created by MOE tool push_codebase.
MOE_MIGRATION=5240


git-svn-id: https://closure-library.googlecode.com/svn/trunk@2057 0b95b8e8-c90f-11de-9d4f-f947ee5921c8