Automated g4 rollback.
*** Reason for rollback ***

It breaks a unit test

*** Original change description ***

Make trodedit bubbles (like the link bubble with 'Go to link... Change...') keyboard accessible. Since the bubble is created in a very different place in the DOM than the target element, this involves capturing the TAB event
This CL is not intended to change the behavior of link bubbles in any products, UNLESS it is turned on explicitly.

R=vbhasin
DELTA=160  (1 added, 153 deleted, 6 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=5574


git-svn-id: https://closure-library.googlecode.com/svn/trunk@2193 0b95b8e8-c90f-11de-9d4f-f947ee5921c8