Allow text templates to be rendered with renderElement and friends.
The output of the templates will be escaped.

This also removes the special case for SanitizedHtmlAttributes which
didn't make sense.  It was previously there because I mistook
HTML_ATTRIBUTE (which was the previous name) for an attribute value,
rather than a set of key/value pairs.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=42858624