Remove the inconsistency between goog.result.waitOnSuccess and goog.result.waitOnError by passing in the error value as the first argument to the callback function.
In the preparation of the CL it was realized that a number of code
sites had their first statements as result.getError(). The use of this
style rather than just passing in just the result in all cases leads to
lesser boilerplate code and has thus been preferred.

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=43075687