Add debug assertions to goog.events to catch the following issues: * using listen/dispatchEvent on un-initialized goog.events.EventTarget * using goog.events.dispatchEvent with non-goog.events.EventTarget instance.
The hope is that this will help catch possible breakages for when we flip
USE_LISTENABLE_INTERFACE by asserting in debug mode (it'll still work in
prod).

Also removes assertions in the new event target implementation that checks
that the instance has not been disposed when listen/dispatchEvent is called.
There are too many tests that depend on this undocumented behavior (it works
with goog.events, it still work with the new implementation, if we remove the
assertions). We can attempt to do this as separate effort -- I can't justify
holding back fixing leaking listeners for this, especially because in the new
world, listening/dispatching on disposed event target is perfectly fine,
leak-wise (it's not today).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=43157289