This CL makes a few changes to IframeIo that allow it to be used for cross-domain form submits.
Specifically, it's meant to be compatible with submitting to
a different domain where the responding server will return code
that postMessages back a response that the code on the original
domain can read.

The primary changes are that the IframeIo optionally doesn't add
event listeners for the form response (since adding those
listeners to a cross-domain iframe is a security violation)
and that it makes the method to get the internal iframe public
so that the calling code can get its window object for
cross domain communication.

This also fixes a minor bug that causes the internal iframe to
create a user visible artifact (scrollbar on the containing
window) in Chrome under some circumstances.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=43282460