Automated g4 rollback of changelist 42432846.
*** Reason for rollback ***

These changes to goog.ui.ac.Renderer.prototype.hiliteRow
are expensive enough to cause a significant lag in highlighting
the elements of the address autocomplete dropdown, especially
when the browser is zoomed (e.g. 110%). The lag is
very noticeable when just moving the cursor around in this
dropdown - sometimes one can move the mouse into a position
and highlight never comes at all until there's another wiggle
of the mouse. We believe this is behind a large number of
user complaints noted in b/8284080 (starting on 2/21),
as the autocomplete code selects the highlighted item
rather than the item being clicked.
(Whether that's the Right Thing is another topic, but the laggy
highlight behavior here is  UI issue even if the right object
is selected upon click).

*** Original change description ***

Fix a row under the mouse cursor beeing selected when rows scroll.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=43471833