Automated g4 rollback
*** Reason for rollback ***

Roll forward the original fix. This version is closer to the
original code, and is explicitly tolerant of an invalid or
missing value for e.row.  Thus, if there's a valid row
specified and it's not currently highlighted, we highlight it
before the inevitable call to selectHilited().

*** Original change description ***

When a SELECT event comes in on an autocomplete row,
hilite (sic) it first so that selectHilited() will actually
select the correct row. This prevents issues where
a mouse event comes in on a row that is not currently the
hilited row, e.g. if the autocomplete list appeared under
the mouse and the user clicks very quickly.

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=44919875