Add goog.dom.iframe.BLANK_SOURCE_NEW_FRAME, an iframe source that prefers to not reload an iframe.
This can be used to minimize browsers that require waiting for the async load of an iframe. Practically, this allows iframes in FF/Gecko to be detected as loaded after being added to the DOM.

NOTE: This constant is not a gurantee and some approach similar to goog.net.IframeLoadMonitor should still be used to detect iframe load.

NOTE: Cross-browser spec compliance for the details of the 'javascript:' url spec varies.

NOTE: This should not be used to replace the source of an existing iframe. goog.dom.iframe.BLANK_SOURCE it appears to be used in places to do that, hence introducing a new constant.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=45630335