Enable vulnerability workaround in Safari publicly now that fix has been out.
Safari has a nasty bug where if you have an http URL with a username, e.g.,
http://evil.com%2F@google.com/
Safari will report that window.location.href is
http://evil.com/google.com/
so that anyone who tries to parse the domain of that URL will get
the wrong domain. We've seen exploits where people use this to trick
Safari into loading resources from evil domains.

To work around this, we run a little "Safari phishing check", and throw
an exception if we see this happening.

There is no convenient place to put this check. We apply it to
anyone doing URI parsing on Webkit. We're not happy about this, but
it fixes the problem.

This should be removed once Safari fixes their bug.

Exploit reported by Masato Kinugawa.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=46998287