Make sure error.message is not undefined.
If error.message is undefined, then the truncation introduced to
ErrorReporter in CL 47061426 causes an exception.

The message can be undefined when error is not an Error object. This can
happen with any object, since anything can be thrown, but one common case
is when error is a GWT exception.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=49119020