Regards PopUpBase as displayed at start of transition, for blur event handling.
Pop-ups can hide automatically on a blur event, but they often receive one as part of being launched. To avoid hiding at that time we ignore blur events for the first 150 ms after showing the pop-up.

We can however also animate the opening of the pop-up, in which case we only fire the show event once the animation completes. For JFK this transition is 218 ms.

Consider the pop-up displayed at the start of the animation, so that we correctly ignore the blur event for the first 150 ms.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=49525411