Add a BoundableCollectableStorage.
Apart from the same functionalities as a CollectableStorage it limits the number of items in a storage.

It keeps a list of keys as an local storage item. Keys in list are in order of creation date (in fact, a date of last 'set' operation for a given key).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=50562704