RFC: XhrLike interface and typedef.
There are multiple ways we can spin it, but I went with this route
since it only required adding one file, but the "OrNative" thing is
weird.

We also considered adding an interface directly to the externs, but we
thought that might be weird.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=53516438