Adds a new transport, DirectTransport, for xpc.
This transport allows you to communicate directly with a peer that is in the same security context.
This is so we can use the same interface to chat with our peer when its on a separate domain and when its on the same domain.
This was implemented for several reasons but one is because Yahoo!'s safe frames implementation blocks postMessage.

This is mainly copied from the NativeMessagingTransport. One things I don't really like is that I also need to support a peer in the same window so I have add the role to the channel name. If you have a better idea about how to handle this let me know.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=54903705