Automated g4 rollback
*** Reason for rollback ***

Rolling forward after hiding the behavior behind an option.

*** Original change description ***

Automated g4 rollback

*** Reason for rollback ***

*** Original change description ***

In the AutoComplete renderer, set overflow options to RESIZE_HEIGHT, so that if the suggestion dropdown is too large for the screen it is resized to fit. Also set "overflow-y: auto" on the dropdown element, so that if it is truncated to fit on the screen, then it gets scrollbars.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=55384618