Add a couple of attributes of <input> elements to DIRECT_ATTRIBUTE_MAP_
'value' is needed because it conflicts with HTMLInputElement.value. Though technically, the right value will show up on the page, it won't happen as a result of setting an attribute on the tag.

'list' is needed because it conflicts with HTMLInputElement.list, which refers to the actual DOM node of the data list. I think the correct behavior here should be for createDom to accept a string, because we're creating HTML.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=55419571