Changes goog.ui.Dialog's decorateInternal() to not read back its content element's .innerHTML.
Doing so appears to have been unnecessary: The read only took place if the
dialog had a content element at the time decorateInternal was called.  In that
case, the element's innerHTML was saved in Dialog's content_ field; however the
value is then never used again (content_ is used in createDom, but that
wouldn't be sensibly called after decorateInternal()).

It is possible for client code to observe the change introduced by this CL:
Client code that uses decorate/decorateInternal to initialize a dialog, and
then reads getContent(). However, this wouldn't seem to make much sense
(except perhaps in tests). A review of (the few) calls to getContent from
production code indicates this pattern is not used in practice, and there were
no test failures on global presubmit.

Why make this change?
* Reading .innerHTML can result in security (XSS) vulnerabilities due to bugs
in browser implementations of HTML reserialization (see
http://dl.acm.org/citation.cfm?id=2516723). This change removes the
possibilty of such bugs due in Dialog's implementation.
* It simplifies a planned refactoring of goog.ui.Dialog to use
goog.html.SafeHtml as the type of its content_ field.
Removing the read of innerHTML avoids the need for an unsafe conversion from
string to HTML in Dialog's implementation.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=56869273