Make goog.ui.ac.AutoComplete always highlight the first row on mouse click.
As can be seen in the newly added regression test case, clicking on the first item of an AutoComplete with autoHilite=false without first dispatching a mouseover event can lead to the clicked row not being selected. This is due to a check of the variable "rowId" which is probably supposed to check that "rowId" is not the empty string but also fails on the value 0.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=58386485