Automated g4 rollback of changelist 61092036.
*** Reason for rollback ***

Broke opening a dialog as a result of a prompt close.

This is because the prompt listens to the Dialog SELECT event, which is
fired before the dialog is hidden. So if anything is focused on the
prompt callback, the focus is lost when the dialog is hidden.

*** Original change description ***

Fix restoring the last focused element on ModalPopup such that:
1) The focus is restored if the popup has a transition.
2) The focus is restored regardless of the current focused element.
This allows the focus to be restored even when the control that
triggered the hide still has focus. This is done before the call
to onHide so subclasses can still override it to focus other
elements.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=61334459