Changes boolean hasKeyBoardControl_ to menuIsVisible_ and sets this whenever setSubMenuVisible is called.
This change is necessary because we can get into a case where the actual menu inside the SubMenu has been closed before the SubMenu has processed the event. This means that when setSubMenuVisible(false) is called the menu is already hidden and we do not set hasKeyBoardControl_ to false as desired. This causes the SubMenu to mishandle the next keyboard event.

To fix this we rename the variable to menuIsVisible_ and set it to the param of setSubMenuVisible whenever that method is called. This variable now represents whether or not the SubMenu thinks it's inner menu is visible or not (which may differ from the visibility set on that menu). We then use that variable to control when we set the goog-submenu-open CSS class, rather than the actual visibility of the menu.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=65627687