Updated fielddescription to always return a new instance every time getDefaultValue is called.
This fixes an issue where getting the default value of a field that is not a native value, a single instance is returned for every call. Modifying this instance will change the value for future calls to getDefaultValue since it's a reference to the internally stored one. This leads to hard to find bugs.

Discussion: https://groups.google.com/a/google.com/forum/#!topic/closure-users/QZGj7TUY0AA
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=66524199