Adds support for "auto hide partners" to goog.ui.PopupBase.
Auto hide partners are other elements that should not cause the
popup to be auto-hidden when when a mouse event occurs within them.
A common example would be a dropdown toggle button tied to a popup.

The use of auto hide partners means you no longer need to rely on
isOrWasRecentlyVisible() and the 150ms heuristic for determining
if clicking on a toggle button should hide or show a popup.

This change should not impact existing users of
isOrWasRecentlyVisible(). It adds addAutoHidePartner(element) and
removeAutoHidePartner(element) as optional configurable options to
the popup.

The API mirrors PopupPanel.java from the Google Web Toolkit.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=66554471