Automated g4 rollback of changelist 66785455.
*** Reason for rollback ***

Caused test failures: https://test.corp.google.com/ui#cl=66785455

*** Original change description ***

Modal popup: Reuse the viewport size to avoid some layout thrashing.
When showing the modal popup, first the viewport size is measured to determine the size of the background and later it's measured again to determine the position to center the popup. In between the size of the background is set. So the second time the viewport size is measured, the browser needs to recalculate styles and layout. We avoid this by avoiding measuring the viewport size for a second time.

See also #13 on b/13672409...

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=66791422