Change goog.events.EventTarget to use .concat rather than goog.array.clone as the instances being cloned are Array and not ArrayLike.
Unscientifically, this should be a small perf win, as microbenchmarks demonstrate concat is faster than the code in goog.array.clone. Profiles in Docs verify this to be the case for cases that generate many events.

A similar change was done to goog.events in http://cl/65954086
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=66796544