goog.editor.Field#focusAndPlaceCursorAtStart: place cursor for IE9+.
goog.editor.Field#focusAndPlaceCursorAtStart has a comment indicating that the cursor should not be placed in Gecko. However, the code is inconsistent with this comment; the cursor is only placed when (goog.editor.BrowserFeature.HAS_IE_RANGES || goog.userAgent.WEBKIT).

This condition excludes IE9+ (which don't fall under HAS_IE_RANGES because they support the W3C Range API). As a result, the cursor is not placed, and several tests that expect the cursor to have been placed fail with an NPE when window.getSelection is unexpectedly empty.

This CL changes the condition to place the cursor in all browsers except IE<9 and Firefox.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=70918574