Add support to the closure graphics library for applying general affine transforms to elements.
We realize this library is deprecated since modern browsers support SVG and Canvas, but we are using this library in the Imagery Viewer (StreetView, photos, and other imagery) to draw vector graphics on older browsers like IE8.  The reason we need to support IE8 is that the Imagery Viewer is being integrated into the JS Maps API v3 for Street View, and the JS Maps API v3 supports browsers back to IE8.  The JS Maps API v3 already has an older Street View renderer, but we are trying to consolidate Street View renderers to improve consistency and the number of viewers (and the amount of code) that we need to maintain.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=72413581