Update proto2 objectserializer to be more robust when deserializing booleans. Specifically, add support for when they are represented as 0/1 rather than true/false.
Motivation:
There are existing libraries which can generate this when using object notation serialization. Specifically, this is present in the recommended serializer for GWT protos, MessageSerializerBuilder.java. When used in GWT client, this will generate the 0/1 representations. (however not when run as native Java on a server)

This change will enable JS applications to use common serialization to communicate in a shared application environment
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=73569780