Automated g4 rollback of changelist 73824704.
*** Reason for rollback ***

e.target is not defined, so this blows as an exception in the code. It could lead to functional bugs.

*** Original change description ***

Adds the ARIA role of tab to the roles for which goog.events.actionEventWrapper will process space in accordance with the ARIA specs.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=73875600