Made _GetInputByPath in closurebuilder.py use os.path.realpath.
A problem occured if 'closurebuilder.py' was run with relative path inputs
and the sources were absolute paths with symlinks within. 'os.path.abspath'
always resolves symlinks in the stem it adds to a relative path even though
it doesn't ever resolve symlinks for absolute paths.

The solution was to use 'os.path.realpath' instead which resolves all
embedded symlinks, providing an apples-to-apples comparison.