Deprecate the (new) goog.string.isEmptyOrWhitespaceSafe function.
The function was added as part of the preamble to the work described in go/lsc-goog-string-isempty, but in the end it was decided not to migrate to this function.  Deprecate it so that it doesn't get adopted too widely for now.  Once the linked LSC is done, a followup CL will be created to fix remaining clients, and to remove this function.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=79106013