When walking up the iframe tree, check that the current window's parent reference doesn't point to itself.
In IE, if an iframe is removed, its window's parent reference is itself (along with .self and .top). This can lead to an infinite loop when trying to get an offset of an element in an iframe that's been removed when the relativeWin isn't the window of said iframe.

Tested:
- blaze test javascript/closure/...
- test failure prior to change: https://sponge.corp.google.com/invocation?id=b268fbc3-869b-41a3-808e-3421b77a4385&searchFor=user%3Ajeshaw
- passing with change: https://sponge.corp.google.com/invocation?id=3d95442f-30c6-406b-a77b-98007743dc54&searchFor=user%3Ajeshaw
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=83265208