Replace use of innerHTML and document.write() with goog.dom.safe. This lets us remove fancywindow.js from the innerHTML whitelist in Conformance.
To support this formatRecordAsHtml() is made part of the Formatter base class, so that both formatters can return SafeHtml.

Resubmission note: This is the third attempt at this CL. The first one, in cl/83033183, was incorrect because it concatenated a SafeStyleSheet with plain strings without unwrapping it first. The second one, in cl/84421558 broke a user of DebugWindow.getStyleRules(), this CL updates the user code.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=84963944