Remove goog.dom.canUseQuerySelector_.
It was added in cl/16436453 to address a bug in Webkit. Its Webkit-related logic was removed in cl/29042474 when we dropped support for the affected versions of that browser. Currently it adds no value and is causing code needing either querySelector or querySelectorAll to check for the existence of both.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=84978219