Automated g4 rollback of changelist 85148725.
*** Reason for rollback ***

Causes internal JSCompiler error ---

*** Original change description ***

Replace use of document.write() with goog.dom.safe. This lets us remove fancywindow.js from the document.write() whitelist in Conformance. debugwindow.js is also removed from whitelist as it was not using document.write().

There's a small difference in the generated HTML. The DOCTYPE is no longer to "transtional". As far as I can tell this should be fine because the only tag that requires transitional is <big> and if it gets ignored it should be fine, given that this is debug UI.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=85168292