Replace use of document.write() with goog.dom.safe. This lets us remove fancywindow.js from the document.write() whitelist in Conformance. debugwindow.js is also removed from whitelist as it was not using document.write().
There's a small difference in the generated HTML. The DOCTYPE is no longer to "transtional". As far as I can tell this should be fine because the only tag that requires transitional is <big> and if it gets ignored it should be fine, given that this is debug UI.

Note: this is the second attempt at this CL. The first one, in cl/85148725, causes an internal JSCompiler error in a project due to the use of goog.scope() (b/19210681).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=85410432