Actually report errors during module callback execution.
In stark contrast to the comment the errors got transformed to a boolean and later to some kind of error code.

This was bad because
- Engineers saw something like: Error code 2
- The errors were never reported to server side logging tools.

With this CL as large source of errors that were previously swallowed will be actually made visible. This should resolve some mysteries of the past.
Because errors are thrown in their own JS execution context no existing behavior is changed.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=86408897