Ensure deterministic results with goog.testing.TestCase Promise-based tests in IE11.
In cases where a browser behavior requires yielding to the browser's
execution thread, an IE11 test may succeed if it is the first test run on
the page, but will fail if it is run later on. This is due to subsequent
tests being scheduled using Promises, while the first test initiates the
execution chain.

This change fixes the issue by executing the first test as a child Promise,
just as subsequent tests are.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=86825653