Accept SafeHtml in addition to string as html in DrilldownRow constructor.
A new unsafeCreate() is introduced. We do this instead of directly using legacyconversions in the constructor because we cannot create a Conformance rule to match correctly on record types passed to the constructor for this case. This is due to how type inference is done by JSCompiler.

This is part of a 3 step process to make DrilldownRow XSS-safe:
1. This change.
2. Refactor users to pass SafeHtml to the constructor or, for non-trivial uses, move them over to unsafeCreate.
3. Disallow string as html in the constructor and replace use of innerHTML with goog.dom.safe.

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=87050165