Workout around for iOS bug where target is not observed in window.open.
iOS in standalone mode disregards "target" in window.open and always opens new URL in the same window. The workout around is to create an "A" element and send a click event to it. Notice that the "A" tag does NOT have to be added to the DOM.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=87993339