closure: don’t use toLowerCase on tag names; use goog.dom.TagName instead
This change deals only with cases where a tag (or node) name was converted
to lower case and modifies them to use goog.dom.TagName instead.

Not converting to lower case (or upper case) does potentially break code
when run in XHTML (which in contrast to HTML uses lower-case tag and node
names) but Closure library does not support XHTML anyway so this should
not be an issue.

Notable exception are calls to goog.dom.tags.isVoidTag which, in it’s
interface, requires tag name to be lower case.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=88259142