Fix tests for IE.
What kept the tests from running was that goog.window.open() was reopening the same window for IE due to target being undefined. To fix this we always set the target from tests. Arguably goog.window.open() is broken (it should probably set a default target of _self) but for now we at least get the tests to run.

The test for > is removed because it seems to cause the Closure test server to bork when IE sends it unencoded. It doesn't add any value since there's also a test for < and an HTML-escaping function would always have to escape < (but not necessarily >).
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=89619443