Remove channelrequest.js from Conformance document.write/innerHTML whitelist.
To do this uncheckedconversions are used. The rationale being that this code seems to be superseded by the one in closure/labs/net/webchannel/channelrequest.js and so risking bugs and performance regressions to use goog.html.SafeHtml.create() is not warranted given that it's unlikely to change and likely to be deprecated.

There is a slight change in behavior since now a URL is being sanitized and input HTML-escaped when interpolated into HTML. This closes any potential (even if unlikely) XSS bugs.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=90803104