Automated g4 rollback
*** Reason for rollback ***

Tests with concurrent MockClock instances can fail if a goog.async.run is used.

*** Original change description ***

When goog.async.run is running with MockClock when want to process the workqueue when MockClock say its time, but when MockClock is no longer in control such as when a "continuation" promise is returned from a TestCase test method we also need to do work.

The real "setTimeout" should never run during a synchronous tes, so whenever we have work we schedule both the "mocked" setImmediate for MockClock (which it pump) and a "real" setTimeout (which the system will pump).  Since it is always valid...

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=93037477