- When the MockClock code is loaded stash a version of "setTimeout" to pass to goog.async.run so that multiple MockClock instances running at the same time dont result in goog.async.run calling a version of MockClock setTimeout on a instance that has been disposed or inactive (calling a MockClock version at all defeats the intent of the original change).
- Add a warning about Mocking "setTimeout" multiple times as the behavior is
undefined (uninstalling MockClock clears any scheduled work) especially when
working with promises such as goog.Promise that depend on setTimeout and
friends to manage their work queues in tests.

This works around the problem discovered with the tests that resulted in the
rollback, gives reasonable feedback until such a time as we can make MockClock
a singleton.

*** Original change description ***

When goog.async.run is running with MockClock when want to process the
workqueue when MockClock say its time, but when MockClock is no longer in
control such as when a "continuation" promise is returned from a TestCase test
method we also need to do work.
The real "setTimeout" should never run during a synchronous tes, so whenever we
have work we schedule both the "mocked" setImmediate for MockClock (which it
pump) and a "real" setTimeout (which the system will pump).  Since it is always
valid to run the workqueue at any entry point this gives us the correct
behavior.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=93044127