Fix js test timeouts by resetting batchTime_ inside setTimeout.
Previously, batchTime_ was reset in finishTestInvocation_, and setTimeout was
called.  If there was a large delay before the timeout function was called, it
would count toward the maxRunTime, producing the need to yield frequently (e.g.
after each test.

The ultimate source of this issue may be an increase in the delay before the setTimeout function is called in Chrome 42 under certain conditions, as the problem appeared on switching from Chrome 41 to Chrome 42.

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=93703014