Conditionally invoke attachEvent() in goog.events.listen_
attachEvent was deprecated in IE9 and removed in IE11, replaced by
addEventListener. It looks like there are some buggy cases in IE11 where
addEventListener doesn't exist. In this case, we're invoking the nonexistant
attachEvent and throwing errors. This bug is currently the top error source for
IE11.

This patch should stop that throw, but not in a particularly satisfactory way.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=94021834