Removing toUsTimeString.
This is something that should never be used.
It formats a time as "12:34 AM" with some small bells and whistles:
* 12/24 hours
* render the minutes or not if the value is 0 ("9 AM" vs "9:00 AM")

Problems: this is nether locale independent (because the AM/PM strings are localized)
nor is it good enough for international use (hard-coded ':', am/pm at the end)

So it can't be used correctly, no matter what.

For communication with a server / another application, one should use something like toIsoTimeString / toIsoString / toUTCIsoString.
For something that will be displayed for a user, one should use goog.i18n.DateTimeFormat

===
But should we first mark it @deprecated for a while? For the outside world to catch up?
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96113003