Fix DST bug where setting hours to 0 in timezones where there is no hour 0 causes the day to go back one. Fixes https://github.com/google/closure-library/issues/34
Writing a test for this behavior is hard because the bug only happens in particular time zones. AFAIK, this would require mocking the native Date constructor and/or methods. At that point, I'd just be testing maybeFixDst_ is working as expected which defeats the purpose of the test altogether.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=99194172