DragListGroup: Allow inherited visibility on dropped items.
When an item is being dragged, its visibility is set to "hidden" and the element is visually replaced by a cloned version elsewhere in the DOM.  Prior to this CL, the visibility was set to "visible" when the item was dropped.  The problem is that visibility:visible elements are drawn on the page even when they are descendants of visibility:hidden elements, so previously dragged elements lose their ability to inherit their parents' visibility.  This is especially a problem for nested drag lists: if you drag and drop a child, and then later drag the parent, the previously dragged children remain visible on the page even as the parent is hidden.  This CL uses the empty string rather than the value "visible" so that visibility logic defers to the style sheet.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=99828941