Merge pull request #424 from elemoine/targettouches
Do not rely on targetTouches for determining "relevant" touch.

targetTouches[0] is unlikely to be the "relevant touch" when there are multiple touches on the touch surface, because, as documented on MDN, targetTouches is the list of "all the Touch objects for touch points that are still in contact with the touch surface". targetTouches[0] may be unrelated to the current touch event, so using its clientX, clientY, screenX and screenY values does not make sense.

changedTouches is the list of "all the Touch objects for touch points that contributed to this touch event", so it makes more sense to use changedTouches[0] as the "relevant touch" when setting the BrowserEvent's clientX, clientY, screenX and screenY values.

See https://github.com/google/closure-library/pull/424 for context.

-------------
Created by MOE: http://code.google.com/p/moe-java
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=100621715