Automated g4 rollback of changelist 100621715.
*** Reason for rollback ***

It has broken some tests:

*** Original change description ***

Merge pull request #424 from elemoine/targettouches

Do not rely on targetTouches for determining "relevant" touch.

targetTouches[0] is unlikely to be the "relevant touch" when there are multiple touches on the touch surface, because, as documented on MDN, targetTouches is the list of "all the Touch objects for touch points that are still in contact with the touch surface". targetTouches[0] may be unrelated to the current touch event, so using its clientX, clientY, screenX and screenY values do...

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=100634635