Set failOnUnreportedAsserts to be false on tests that fail with it true.
failUnUnreportedAsserts is a new flag that causes assertions in Google’s JavaScript tests to be treated the same as in most every other testing framework: high-severity exceptions that cannot be unintentionally suppressed. To implement this change without breaking existing tests, we are explicitly setting this flag to be false on tests that would fail with it true.

Tested:
    TAP train for global presubmit queue
    http://test/OCL:110372922:BASE:110384772:1450324356300:7ab073b3
    Some tests failed; test failures are believed to be unrelated to this CL
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=110481746