Prevent calling enterDocument twice for popupDatePicker_ child of InputDatePicker.
The superclass call for enterDocument ends up calling enterDocument for popupDatePicker_ because it was added as a child in the InputDatePicker constructor.  The explicit call to this.popupDatePicker_.enterDocument() is not necessary (and causes popupDatePicker_.datePicker_.decorate() to run twice, which shows 2 calendars when the popup opens) so this CL removes it.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=101118302