Wait for the Worker to start up before running the PortChannel tests, and give it a very generous amount of time to start.
On some machines, the first test in alphabetical order (in this case, testPort) would fail because, after 3 seconds, the Worker still hadn't started. This resulted in a 40% failure rate for this test on safari-osx, and a 100% failure rate on Chrome on my workstation.

Also, re-enable testWindow in Safari. It was disabled 3 years ago by CL 31139964 for reasons that are no longer applicable.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=102070939