Handle Multiple Number Groupings (resolves https://github.com/google/closure-library/issues/174)
Applied the feedback from @joeltine including style changes and breaking down the function into smaller functions

Include only 1 name in Authors; take in comestic changes from @joeltine

Fixed up to meet up to the Google style guide and standards. Thanks to @joeltine.

Fixed up to meet up to the Google style guide and standards. Thanks to @joeltine.

Removed new line in the test. Thanks to @joeltine.

Made a number of changes from @joeltine 's feedback.

Made a number of changes from @joeltine 's feedback.

Fixed a type mismatch due to what the functions are returning based on @joeltine 's feedback.

Removed accidental file

Changed the properties of the parameters in the JSDOc to recognize what is expected to be null or not null

Changed the descriptions about the variables that is being passed into the smaller functions