Never publish to a listener after it has been unsubscribed, by converting the listener to a no-op internally.
Without this behavior, listeners can be called after unsubscribe depending on the state of the stack. For example, subscribeOnce does not actually subscribe once if it is triggered while publishing. Any code that can publish events in a non-nested or nested way will have unexpected behavior. Unsubscribing immediately is the norm for JS, as it is specified in the w3c spec for the DOM events.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=103954434