Remove iframe_xhr_test.
This test was only run on Gecko browsers to verify a workaround for Firefox 2.
A comment left in 2011 says the original reason for this code no longer exists, but kept the test to track regressions.
Since this test was written in a unique style and based on the deprecated AsyncTestCase, it seems like a fine time to consider it obsolete.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105618704