Convert IframeLoadMonitorTest and MultiIframeLoadMonitorTest from AsyncTestCase to Promises.
Fixes inconsistent behavior in IframeLoadMonitorTest: the hasContent flag was
added to work around inconsistent loading behavior between 2008-era versions of
IE and Firefox. Firefox would fire one IFRAME load event as soon as an IFRAME
was added to the document (presumably defaulting to about:blank), and then again
once the actual SRC had loaded. The workaround was to test for the presence of
content below the document body, but only for non-IE browsers.

Updated the code to consistently check for document.body.firstChild in all
browsers (when requested with the hasContent flag). Updated tests to exercise
this behavior, and to actually fail when the "hasContent" flag is flipped in
either test.

(It's a good question whether these classes are still necessary given
improvements in load event handling, but I'll leave deprecation discussions for
the future.)
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=106001462