Deprecate guardBracketInHtml.
This function is broken, it corrupts for example this input: ( <br title=")">
Output: <span dir=ltr>( <br title=")</span>">

It also works with HTML stored in strings which is unsafe and which should be evicted in favor of SafeHtml.

I would also like to see a hard evidence that such a function is still needed. The description is very vague: "problem of messy bracket display, which frequently happens in RTL layout". Where does this problem happen? Maybe it's only in browser versions that are not supported anymore?

A future CL will delete this function.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=107429496