Make goog.testing.dom.assertHtmlContentsMatch output more accurate "expected" string.
Currently it outputs the given string, instead of reading the div's innerHTML which is what it actually uses for comparison. On the other hand it does use innerHTML for the "actual" pattern.

This makes the two strings look different, and even contain different attributes (when running with Polymer, since it modifies the DOM), which makes finding the real differences between the actual and expected HTML very difficult for the developer.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=107561115