use &lt; instead of < in comment
In goog.events.eventid the comment has sample code which includes text
that looks like an HTML tag.  The text is:
    /** @type {goog.events.EventId<MyEventObj>}
When building documentation with js-dossier, this comment generates
warnings:
    [WARNING][com.github.jsdossier.soy.HtmlSanitizer]
    Discarded tag "myeventobj" in text:

Change to use &lt; instead of < in that comment to prevent these
warnings.  Additionally, use <pre> tags so that the comment appears
as a comment block with the original line breaks (instead of being
run together into a long line of text).  Further, add a zero-width
space character so that the closing comment symbol */ can be added
to the sample code.