Merge pull request #603 from myphysicslab/closure-library
use markdown code block in comment of eventid

In `goog.events.eventid` the comment has sample code which includes text
that looks like an HTML tag.  The text is:

     /** @type {goog.events.EventId<MyEventObj>}

When building documentation with js-dossier, this comment generates
warnings:

     [WARNING][com.github.jsdossier.soy.HtmlSanitizer]
     Discarded tag "myeventobj" in text:

Change to use a markdown "4 spaces indent" code block in that comment
to prevent these warnings. This also preserves the line breaks in the
code sample. Further, add the closing comment symbol *\ to the code
sample (with a backslash so it doesn't actually close the comment
block).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=108384222