Fix a bug where vec3 slerp would sometimes produce NaN results.
The dot product of the two vectors was sometimes very slightly greater
than 1, or less than -1, both of which produce a NaN when passed into
Math.acos().

Fixes this by deciding whether the vectors are the same/opposite directions
based on the dot product result itself and only using the acos() to compute
the angle if the vectors are not in the same/opposite directions (ensuring
the argument we pass in is in [-1, 1]).
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=108274541