Ignore whitespace in base64.decodeString, and tolerate webSafe encoding Remove opt_webSafe param from the decodeStringToXxxArray methods
The opt_webSafe optional parameter is misleading and inconsistent in behavior:
See https://github.com/google/closure-library/issues/600

IE is non-compliant and barfs if passed whitespace in atob() (but note Edge
is compliant)

Given the added complexities in using atob(), the simplest path to a
consistent behavior would be remove the use of atob() and make decodeString()
unconditionally call through to the portable implementation. However that would
regress perf by around a factor of x10 on Chrome, FF so instead just blacklist
implementations known to be non-compliant.

Note that prior to this CL charToByteMap_ and charToByteMapWebSafe_ had identical
content, so the removal of the latter is only simplification, not a change in
observable behavior.

Bug: https://github.com/google/closure-library/issues/600
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=109725759