Ensure ScrollFloater correctly handles horizontal scrolling.
When a scroll event is detected and the ScrollFloater element is already floating, update the left fixed position based on the viewport's horizontal position.

The crux of this issue is that when the window has a horizontal scrollbar, the act of 'floating' the scrollbar moves it to a fixed position based on the _absolute_ page position of the scrollbar. So if you have a small window with a scrollfloater and are scrolled to the right, the scrollfloater will jump to the right when you scroll down, and will jump back to the left when you scroll up and repin the element. This CL ensures that the element maintains it's left alignment no matter what state the element is in, and wherever the user is scrolled in their window.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=110037788