Add charset in all HTML test files.
When using js files with UTF-8 encoding and loading those js files from an HTML file, without specifying the charset, the browser will try to guess the encoding, which might end up with the wrong assumption.

This triggers interesting bugs where the same test pass or not depending on the machine or the condition when the test was run (alone, part of a test suite etc...).

This CL brought to you by
$ for i in $(find . -name "*_test.html" |xargs grep -L -E "<meta [^>]*charset" ); do g4 edit -c 109908655 $i; sed -i -r 's!(\s*)<title>!\1<meta charset="UTF-8" />\n\1<title>!' $i; done
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=110230739