Fix unreported assertion error in labs/net/xhr_test.js.
Canceling the Promise returned from XHR doesn't yield an XHR error, but a Promise cancellation, so the previous test was erroneously checking for values that did not exist (and passing anyway).
Also removes testAbortRequest because it had exactly the same structure as testCancelRequest.
While we're in there, changes mocked XHR sends to always fulfill a request asynchronously (though possibly with the help of the MockClock).
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=110558386