Fix safari-osx tests for base64_test, flash_test, and loggerclient_test.js to re-eanble Safari in
base64_test: Safari's version of atob now throws when base64 has whitespace. This is a regression and removed Safari from the "NATIVE_SUPPORT" whitelist as a result.

loggerclient_test: This test wasn't properly mimicking normalizeErrorObject for Safari (i.e., looking for "line" and "sourceURL").. which they must've just enabled for "Error()" construction.

flash_test: Safari no longer has "Shockwave Flash" in navigator.plugins, so it was falling through to the navigator.mimeTypes check which ultimately was setting "detectedFlash_" to undefined (mimeType is undefined for Safari). This was a bug in the original implementation... it should always be either true or false.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=111549396